Re-work comparison operators to use composition
This changes the operator implementations to use a `IntPredicate` that
indicates if the comparisons result should be considered a match or not
instead of having an abstract `compare` method that is overriden by
several sub-classes.

The motivation for this change is to be able to re-use the
`IntPredicate` implementations later on in a query optimization for
`array_length` / `array_upper`.